+++ domain/src/commonMain/kotlin/dev/kevinsalazar/exchange/domain/payload/request/DepositRequest.kt
@@ -5,5 +5,5 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class DepositRequest(
     val currencyCode: String,
-    val amount: Float
+    val amount: Double
 )
